\name{confint}
\docType{methods}
\alias{confint}
\alias{confint,pim-method}

\title{Compute confindence intervals of \code{pim} model terms.}

\description{

Computes Wald-based confidence intervals for the terms of a PIM model.

}

\section{Methods}{

\describe{
\item{confint}{\code{signature(object = "pim", parm,level =0.95,...)}:
Returns estimate and confidence intervals for specified \code{parm}. Default is to return all params of the model.
  }
 }
}


\author{S. Kovalchik \email{s.a.kovalchik@gmail.com}}

\keyword{methods}
