\name{fits}
\docType{methods}
\alias{fits}
\alias{fits,anoint.fit-method}

\title{Extract fits from \code{anoint.fit} object}

\description{
Extract the specified \code{anoint} model from a \code{anoint.fit} object.
}

\section{Methods}{

\describe{

\item{fits}{\code{signature(object = "anoint.fit",type=c("obo","uim","pim.exact","pim.approx")}:
Extracts the specified fitted object from a \code{anoint.fit}.
  }
  }
}



\author{S. Kovalchik \email{s.a.kovalchik@gmail.com}}

\keyword{methods}
