\name{anoint-package}
\alias{anoint-package}

\docType{package}
\title{
Analysis of interactions for generalized linear models (GLM) or Cox proportional hazards regression models.
}
\description{
Testing procedures and models for investigating the heterogeneity of treatment effect in a clinical trial with multiple baseline covariates.
}
\details{
\tabular{ll}{
Package: \tab anoint\cr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2015-7-10\cr
License: \tab GPL
}
}
\author{Ravi Varadhan <rvaradhan@jhmi.edu> and Stephanie Kovalchik (maintainer) <s.a.kovalchik@gmail.com>}


\references{
Follmann DA, Proschan MA. A multivariate test of interaction for use in clinical trials. \emph{Biometrics} 1999; 55(4):1151-1155
}

\keyword{ package }
