\name{print}
\docType{methods}
\alias{print}
\alias{print,pim-method}

\title{Print coefficients of \code{pim}}

\description{

Prints the regression coefficients of the fit of a \code{pim}.

}

\section{Methods}{

\describe{

\item{print}{\code{signature(x = "pim",digits=4,...)}:
Prints call, coefficients, and responsiveness parameter for \code{pim} object.
   }
  }
}



\author{S. Kovalchik \email{s.a.kovalchik@gmail.com}}

\keyword{methods}
