\name{prognostic.score}
\docType{methods}
\alias{prognostic.score}
\alias{prognostic.score,pim-method}

\title{Prognostic scores for \code{pim} object.}

\description{

Computes prognostic score from \code{pim} model, which is the linear predictor for the main effects of the prognostic factors for untreated subjects.

}

\section{Methods}{
\describe{

 \item{prognostic.score}{\code{signature(object = "pim")}:
Prognostic score at baseline.
  }

 }
}


\author{S. Kovalchik \email{s.a.kovalchik@gmail.com}}

\keyword{methods}
