\name{show.formula.anoint}
\docType{methods}
\alias{show,formula.anoint-method}

\title{Show \code{formula.anoint} object}

\description{

Returns the formula of the \code{formula.anoint} object.

}

\section{Methods}{

\describe{

\item{show}{\code{signature(object = "formula.anoint")}:
Shows formula of the \code{formula.anoint} object.
   }
  }
}



\author{S. Kovalchik \email{s.a.kovalchik@gmail.com}}

\keyword{methods}
