\name{show.anoint}
\docType{methods}
\alias{show,anoint-method}

\title{Show \code{anoint} object}

\description{

Shows the formula specification of the \code{anoint} object.

}

\section{Methods}{

\describe{

\item{show}{\code{signature(object = "anoint")}:
Shows formula call for \code{anoint} object.
   }
  }
}



\author{S. Kovalchik \email{s.a.kovalchik@gmail.com}}

\keyword{methods}
