\name{show-anoint.fit}
\docType{methods}
\alias{show,anoint.fit-method}

\title{Show table of LRT global test results for \code{anoint.fit} object}

\description{

Prints a matrix with logical indicators of whether the given \code{anoint} test rejected the null hypothesis for the global test of interaction.
}

\section{Methods}{

\describe{

\item{show}{\code{signature(object = "anoint.fit")}:
Matrix of logical indicators for global LRT test rejection.
   }
  }
}



\author{S. Kovalchik \email{s.a.kovalchik@gmail.com}}

\keyword{methods}
