\name{show}
\docType{methods}
\alias{show}
\alias{show,pim-method}

\title{Show coefficients of \code{pim}}

\description{

Shows the regression coefficients of the fit of a \code{pim}.

}

\section{Methods}{

\describe{

\item{show}{\code{signature(object = "pim")}:
Shows call, coefficients, and multiple interaction parameter for \code{pim} object.
   }
  }
}



\author{S. Kovalchik \email{s.a.kovalchik@gmail.com}}

\keyword{methods}
