\name{summary}
\docType{methods}
\alias{summary}
\alias{summary,anoint-method}

\title{Summary of \code{anoint} model fit.}

\description{

A list of estimates and convergence status of a \code{anoint} model fit. When \code{exact} is \code{TRUE} the test statistic is the likelihood ratio test, otherwise the Follmann approximate likelihood ratio test statistic is used.

}

\section{Methods}{

\describe{
\item{summary}{\code{signature(x = "anoint",...)}:
Calls summary method of underlying model fit.
  }
 }
}


\author{S. Kovalchik \email{s.a.kovalchik@gmail.com}}

\keyword{methods}
